Move /src/api/helpers work to src/helpers
