+++ src/api/create-env-test-suite/controller.js
@@ -3,12 +3,12 @@ import { isNil } from 'lodash'
 import { config } from '~/src/config'
 import { envTestSuiteValidation } from '~/src/api/create-env-test-suite/helpers/schema/env-test-suite-validation'
 import { updateOverallStatus } from '~/src/api/create-microservice/helpers/save-status'
-import { raiseInfraPullRequest } from '~/src/api/helpers/create/raise-infra-pull-request'
+import { raiseInfraPullRequest } from '~/src/helpers/create/raise-infra-pull-request'
 import { testRunnerEnvironments } from '~/src/config/test-runner-environments'
-import { createTestSuiteStatus } from '~/src/api/helpers/create/create-test-suite-status'
+import { createTestSuiteStatus } from '~/src/helpers/create/create-test-suite-status'
 import { creations } from '~/src/constants/creations'
-import { createTestSuiteFromTemplate } from '~/src/api/helpers/create/create-test-suite-from-template'
-import { createSquidConfig } from '~/src/api/helpers/create/create-squid-config'
+import { createTestSuiteFromTemplate } from '~/src/helpers/create/create-test-suite-from-template'
+import { createSquidConfig } from '~/src/helpers/create/create-squid-config'
 
 const createEnvTestSuiteController = {
   options: {
